% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\docType{data}
\name{emoji}
\alias{emoji}
\title{Emoji}
\format{
\subsection{\code{emoji}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{emoji}{Symbol}
\item{aliases}{Name of the emoji that can be used in a \code{tag}}
\item{tags}{Alternative names}
\item{category}{Category}
\item{description}{Description}
\item{unicode_version}{Unicode Version}
...
}
}
}
\source{
\url{https://github.com/binwiederhier/ntfy/blob/main/web/src/app/emojis.js}
}
\usage{
emoji
}
\description{
Emoji symbols available in ntfy
}
\keyword{datasets}
