% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{get_image_path}
\alias{get_image_path}
\title{Determine filename of a given image file or ggplot object}
\usage{
get_image_path(image)
}
\description{
Determine filename of a given image file or ggplot object
}
\keyword{internal}
