% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy-package.R
\docType{package}
\name{ntfy-package}
\alias{ntfy}
\alias{ntfy-package}
\title{ntfy: Lightweight Wrapper to the 'ntfy.sh' Service}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'ntfy' (pronounce: notify) service is a simple HTTP-based pub-sub notification service. It allows you to send notifications to your phone or desktop via scripts from any computer, entirely without signup, cost or setup. It's also open source if you want to run your own. Visit \url{https://ntfy.sh} for more details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonocarroll/ntfy}
  \item Report bugs at \url{https://github.com/jonocarroll/ntfy/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Carroll \email{rpkg@jcarroll.com.au} (\href{https://orcid.org/0000-0002-1404-5264}{ORCID})

Other contributors:
\itemize{
  \item Andrew Heiss \email{andrew@andrewheiss.com} (\href{https://orcid.org/0000-0002-3948-3914}{ORCID}) [contributor]
}

}
\keyword{internal}
