% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{ntfy_auth}
\alias{ntfy_auth}
\title{Get the ntfy authorization indicator}
\usage{
ntfy_auth(var = "NTFY_AUTH")
}
\arguments{
\item{var}{environment variable in which the ntfy authorization indicator is stored}
}
\value{
a logical that indicates if password authorization is used
}
\description{
Get the ntfy authorization indicator
}
