% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{ntfy_authorization}
\alias{ntfy_authorization}
\title{Create HTTP headers for basic authentication}
\usage{
ntfy_authorization(
  var_username = "NTFY_USERNAME",
  var_password = "NTFY_PASSWORD"
)
}
\arguments{
\item{var_username}{environment variable in which the ntfy username is stored}

\item{var_password}{environment variable in which the ntfy password is stored}
}
\value{
an object with class \code{request} that is passed to \code{\link[httr:POST]{httr::POST()}}
}
\description{
Create HTTP headers for basic authentication
}
\examples{
\dontrun{
Sys.setenv("NTFY_USERNAME" = "example")
Sys.setenv("NTFY_PASSWORD" = "super-secret-password")

# This generates the correct header:
ntfy_authorization()
#> <request>
#> Headers:
#>   * Authorization: Basic ZXhhbXBsZTpzdXBlci1zZWNyZXQtcGFzc3dvcmQ
}

}
