% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{ntfy_done}
\alias{ntfy_done}
\title{Notify Completion of a Process}
\usage{
ntfy_done(
  x,
  topic = ntfy_topic(),
  message = paste0("Process completed at ", Sys.time()),
  title = "ntfy_done()",
  tags = "white_check_mark",
  server = ntfy_server(),
  ...
)
}
\arguments{
\item{x}{a result (ignored)}

\item{topic}{subscribed topic to which to send notification}

\item{message}{text to send as notification}

\item{title}{title of notification. See \url{https://docs.ntfy.sh/publish/#message-title}}

\item{tags}{text tags or emoji shortcodes from \url{https://docs.ntfy.sh/emojis/},
provided as a list}

\item{server}{ntfy server}

\item{...}{other options passed to \code{\link[httr:POST]{httr::POST()}}}
}
\value{
the input x (for further piping) plus a notification will be sent
}
\description{
Notify Completion of a Process
}
