% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{ntfy_done_with_timing}
\alias{ntfy_done_with_timing}
\title{Notify Completion of a Process with Timing}
\usage{
ntfy_done_with_timing(
  x,
  topic = ntfy_topic(),
  body = paste0("Process completed in ", time_result, "s"),
  server = ntfy_server(),
  ...
)
}
\arguments{
\item{x}{expression to be evaluated and timed}

\item{topic}{subscribed topic to which to send notification}

\item{body}{text to send as notification}

\item{server}{ntfy server}

\item{...}{other options passed to \code{\link[httr:POST]{httr::POST()}}}
}
\value{
the result of evaluating x (for further piping) plus a notification will be sent
}
\description{
Notify Completion of a Process with Timing
}
