% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{ntfy_password}
\alias{ntfy_password}
\title{Get the ntfy password}
\usage{
ntfy_password(var = "NTFY_PASSWORD")
}
\arguments{
\item{var}{environment variable in which the password is stored}
}
\value{
the password for the username with access to the protected ntfy topic
}
\description{
Get the ntfy password
}
