% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{ntfy_server}
\alias{ntfy_server}
\title{Get the ntfy server}
\usage{
ntfy_server(var = "NTFY_SERVER")
}
\arguments{
\item{var}{environment variable in which the server URL is stored}
}
\value{
the ntfy server URL
}
\description{
Get the ntfy server
}
