% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{ntfy_username}
\alias{ntfy_username}
\title{Get the ntfy username}
\usage{
ntfy_username(var = "NTFY_USERNAME")
}
\arguments{
\item{var}{environment variable in which the username is stored}
}
\value{
the username with access to the protected ntfy topic
}
\description{
Get the ntfy username
}
