% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{req_add_auth_if_needed}
\alias{req_add_auth_if_needed}
\title{Add basic authorization headers if \code{auth = TRUE}}
\usage{
req_add_auth_if_needed(req, auth, username, password)
}
\description{
Add basic authorization headers if \code{auth = TRUE}
}
\keyword{internal}
