% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{req_add_image_if_needed}
\alias{req_add_image_if_needed}
\title{Add image to the request body if \code{image} is present}
\usage{
req_add_image_if_needed(req, image)
}
\description{
Add image to the request body if \code{image} is present
}
\keyword{internal}
