% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{show_emoji}
\alias{show_emoji}
\title{Show an emoji symbol, or find one by name}
\usage{
show_emoji(name = NULL, search = FALSE)
}
\arguments{
\item{name}{name of emoji to either print or find}

\item{search}{search the \code{tags} column for this name?}
}
\value{
nothing, just prints the emoji if one or more are found
}
\description{
Show an emoji symbol, or find one by name
}
\details{
Emoji are loaded with \code{data("emoji")} and the \code{aliases} column
contains the names compatible with ntfy. Alternative names are
included in the \code{tags} column and these will be searched if the
name is not found in \code{aliases}.
}
\examples{
show_emoji("dog")
show_emoji("party")

}
