% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\docType{data}
\name{tags}
\alias{tags}
\title{Emoji tags}
\format{
An object of class \code{list} of length 1855.
}
\usage{
tags
}
\description{
Emoji symbols compatible with ntfy to be used as tags
}
\keyword{datasets}
