% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{ntfy_send}
\alias{ntfy_send}
\title{Send a Notification}
\usage{
ntfy_send(
  body = "test",
  topic = ntfy_topic(),
  server = ntfy_server(),
  capture = shQuote,
  ...
)
}
\arguments{
\item{body}{text to send as notification}

\item{topic}{subscribed topic to which to send notification}

\item{server}{ntfy server}

\item{capture}{function to use to capture the output, e.g. \code{shQuote}, \code{capture.output}}

\item{...}{other options passed to \code{\link[httr:POST]{httr::POST()}}}
}
\value{
the returned value from \code{\link[httr:POST]{httr::POST()}}
}
\description{
Send a Notification
}
