% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{ntfy_topic}
\alias{ntfy_topic}
\title{Get the ntfy topic}
\usage{
ntfy_topic(var = "NTFY_TOPIC")
}
\arguments{
\item{var}{environment variable in which the topic is stored}
}
\value{
the ntfy topic to which the user should be subscribed
}
\description{
Get the ntfy topic
}
